[FLINK-7520][network] let our Buffer class extend from netty's buffer class
For this, use a common (flink) Buffer interface and an implementation
(NetworkBuffer) that implements netty's buffer methods as well. In the future,
with this, we are able to avoid unnecessary buffer copies when handing buffers
over to netty while keeping our MemorySegment logic and configuration.

For the netty-specific part, the NetworkBuffer also requires a ByteBuf allocator
which is otherwise not needed in out use cases, so if the buffer is handed over
to netty, it requires a byte buffer allocator to be set.