[FLINK-19312][network] Introduce BufferWritingResultPartition to handle ResultSubpartition-related logic (eager partitioning)
In the current abstraction, buffers are directly written to, and read from, ResultSubpartitions, meaning there
is a hard-writed assumption that data needs to be written eagerly partition in the way it will be read.

This change enables other implementations, where records are appended to a joint structure, from which the data
is drawn after the write phase is finished. An example for that would be a sort-based partitioning, which clusters
data belonging to different channels by sorting channel index.

The change migrates the buffer-oriented writing from ResultPartition to the new BufferWritingResultPartition, after
which the interface of ResultPartition free of assumptions that buffers are written directly to a ResultSubpartition.

The resulting interface writes to a joint structure (ResultPartition) and reads from multiple partitioned structures
(ResultSubpartitionView).